+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [1.0.13+1]
 
 - Fixed a bug related to `stepsYAxis` value
+- new property added `displayDataPointWhenNoValue` for `BezierChartConfig`
 
 ## [1.0.13]